+++ django_typer/tests/apps/test_app/management/commands/handle_as_init.py
@@ -1,5 +1,5 @@
 from click import get_current_context
-
+import typing as t
 from django_typer import TyperCommand, command, initialize