+++ examples/polls/models.py
@@ -4,6 +4,7 @@ from django.db import models
 class Question(models.Model):
     question_text = models.CharField(max_length=200)
     pub_date = models.DateTimeField("date published")
+    opened = models.BooleanField(default=True)
 
     def __str__(self):
         return self.question_text